explicity cast to int

+++ small_paino/renderer.cpp
@@ -38,10 +38,10 @@ void draw_rect(float x, float y, float half_size_x, float half_size_y, unsigned 
     x += render_state.width / 2.f;
     y += render_state.height / 2.f;
 
-    int x0 = x - half_size_x;
-    int x1 = x + half_size_x;
-    int y0 = y - half_size_y;
-    int y1 = y + half_size_y;
+    int x0 = (int) (x - half_size_x);
+    int x1 = (int) (x + half_size_x);
+    int y0 = (int) (y - half_size_y);
+    int y1 = (int) (y + half_size_y);
 
     draw_rect_in_pixels(x0, y0, x1, y1, color);
 }
